Changes from PR review moved the Upsert User API endpoint to employerProfilesOuter.API
