+++ src/EmployerProfiles/SFA.DAS.EmployerProfiles/Application/AccountUsers/Commands/UpsertEmployer/UpsertAccountCommandResult.cs
@@ -1,4 +1,4 @@
-﻿namespace SFA.DAS.ApimDeveloper.Application.EmployerAccounts.Commands.UpsertEmployer
+﻿namespace SFA.DAS.EmployerProfiles.Application.AccountUsers.Commands.UpsertEmployer
 {
     public class UpsertAccountCommandResult
     {
